requirements.txt
aiogram>=3.4.0,<4.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
groq>=0.4.0
gradio>=4.0.0